Реализованы CRUD операции

+++ LearnRouting.SimpleMinimalAPI/Models/Employee.cs
@@ -5,9 +5,9 @@ public class Employee
     public int Id { get; set; }
     public string Name { get; set; }
     public string Position { get; set; }
-    public double Salary { get; set; }
+    public decimal Salary { get; set; }
 
-    public Employee(int id, string name, string position, double salary)
+    public Employee(int id, string name, string position, decimal salary)
     {
         Id = id;
         Name = name;
